Move all type extentions to /ext directory files
Move extentions of std::nullptr_t and std::nullopt_t into separate files

This makes able of extention of these types optional in the future.
To make it possible the changes of is_null are needed - overload of
is_null() function depends on include headers order which is not
acceptable. Thats why the implementation has been changed to struct
template overloading.

Move std::vector and std::list extention to separate files
