+++ include/ozo/binary_query.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <ozo/ext/std.h>
+#include <ozo/ext/boost.h>
 #include <ozo/send.h>
 #include <ozo/detail/array.h>
 #include <ozo/concept.h>
