+++ include/ozo/ext/boost.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <ozo/ext/boost/optional.h>
+#include <ozo/ext/boost/scoped_ptr.h>
+#include <ozo/ext/boost/shared_ptr.h>
+#include <ozo/ext/boost/weak_ptr.h>
