+++ include/ozo/ext/boost/scoped_ptr.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include <ozo/type_traits.h>
+#include <boost/scoped_ptr.hpp>
+
+namespace ozo {
+
+template <typename T>
+struct is_nullable<boost::scoped_ptr<T>> : std::true_type {};
+
+template <typename T>
+struct allocate_nullable_impl<boost::scoped_ptr<T>> {
+    template <typename Alloc>
+    static void apply(boost::scoped_ptr<T>& out, const Alloc&) {
+        out.reset(new T{});
+    }
+};
+
+} // namespace ozo
