+++ include/ozo/ext/std.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <ozo/ext/std/list.h>
+#include <ozo/ext/std/nullopt_t.h>
+#include <ozo/ext/std/nullptr_t.h>
+#include <ozo/ext/std/optional.h>
+#include <ozo/ext/std/shared_ptr.h>
+#include <ozo/ext/std/unique_ptr.h>
+#include <ozo/ext/std/vector.h>
+#include <ozo/ext/std/weak_ptr.h>
