+++ include/ozo/recv.h
@@ -5,6 +5,7 @@
 #include <ozo/type_traits.h>
 #include <ozo/size_of.h>
 #include <ozo/concept.h>
+#include <ozo/ext/std.h>
 #include <ozo/detail/endian.h>
 #include <ozo/detail/float.h>
 #include <ozo/istream.h>
